requirements.txt
pyrebase4
fastapi
uvicorn
python-dotenv